fix G

+++ src/dockerserverman.cpp
@@ -316,8 +316,8 @@ bool CDockerServerman::CheckAndCreateServiveSpec(DockerCreateService createServi
         spec.taskTemplate.containerSpec.image = "massgrid/10.0-base-ubuntu16.04";
     spec.taskTemplate.containerSpec.user= "root";
     
-    spec.taskTemplate.resources.limits.memoryBytes = serviceItem.mem.Count *G;
-    spec.taskTemplate.resources.limits.nanoCPUs = serviceItem.cpu.Count * G;
+    spec.taskTemplate.resources.limits.memoryBytes = (int64_t)serviceItem.mem.Count *G;
+    spec.taskTemplate.resources.limits.nanoCPUs = (int64_t)serviceItem.cpu.Count * G;
 
     Config::GenericResources otherresources;
     otherresources.discreateResourceSpec.kind = serviceItem.gpu.Name;
